requirements.txt
buttplug-py
Pillow